docsubtype: https://github.com/metanorma/annotated-express/issues/46
